refactor:설정 페이지를 SSR / CSR 컴포넌트로 분리-#116
